test(cli): move env-origin regressions to the integration boundary
The new `.env`-origin regressions wrote temp files and mutated process
environment variables inside `cli/commands/dev/dev.test.ts` and
`cli/mcp/remote-file-tools.test.ts`, both declared hermetic units. The
semantic unit-boundary audit rejected the undeclared filesystem-write and
process effects.

Relocate those tests to
`tests/integration/semantic-unit-boundary/cli/...`, matching the
destination the migration inventory already records for these files, and
stub fetch through `withMockFetch` so the outbound transport moves with the
global. The unit files return to their state on main; coverage is preserved
because the integration:legacy-tests-root suite runs everything under
`tests/`.

The same run also converts the new project-creation regression to
`makeTempDir` and `withMockFetch` so it stops raising the testing
front-door ratchet.
